user-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
Werkzeug==2.3.6
